7agat kteeer

+++ reliable-plumbing-client/src/app/user-management/system-users-management/system-users-management.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { UserManagementService } from '../services/user-management.service';
+import { EnvironmentService } from '../../services/environment.service';
 import { NgbModal, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
 import { RegistrationMode } from '../../models/enums';
+import { NotificationService } from '../../services/notification.service';
 
 @Component({
   selector: 'rb-system-users-management',
@@ -14,7 +16,8 @@ export class SystemUsersManagementComponent implements OnInit {
   registerModelRef: NgbModalRef;
   registrationMode: RegistrationMode = RegistrationMode.admin;
 
-  constructor(private userManagementService: UserManagementService, private modalService: NgbModal) { }
+  constructor(private userManagementService: UserManagementService, private modalService: NgbModal,
+    private environmentService: EnvironmentService, private notificationSerive: NotificationService) { }
 
   ngOnInit() {
     this.getAllSystemUsers();
@@ -23,6 +26,8 @@ export class SystemUsersManagementComponent implements OnInit {
   getAllSystemUsers() {
     this.userManagementService.getAllSystemUsers().subscribe(results => {
       this.users = results;
+      let currentUser = this.environmentService.currentUser;
+      this.users = this.users.filter(user => user.id != currentUser.id)
       for (let user of this.users) {
         user.rolesString = '';
         for (let i = 0; i < user.rolesObj.length; i++) {
@@ -46,8 +51,14 @@ export class SystemUsersManagementComponent implements OnInit {
   }
 
   deleteUser(user) {
-    debugger;
-    this.userManagementService.deleteUserById(user.id).subscribe(success => {
+    let message = 'Are you sure you want delete ' + user.firstName + ' ' + user.lastName;
+    this.notificationSerive.confirmDialog(message, () => {
+      this.userManagementService.deleteUserById(user.id).subscribe(success => {
+        if (success) {
+          this.users = this.users.filter(u => user.id != u.id);
+          this.notificationSerive.printSuccessMessage('User deleted succefully');
+        }
+      });
 
     })
   }
